[day01][TIL] part1: dict comprehension yaay!
Today I learned/implemented:

Braces `{}` can be used for dictionary and set comprehension (inline
definition of collections), same thing works with `[]` for lists, and
with `()` for generator comprehension.

Generator comprehensions are very cool because they do a similar thing
as list comprehension, but they do not generate a list object, meaning
that memory is no allocated for it. Very, very, cool stuff!

More [here](https://www.geeksforgeeks.org/comprehensions-in-python/)