[day01][TIL] enter `enumerate`!!
Today I learned that instead of using indices, there is a much nicer,
built-in way of keeping track of collection iterators (if that's what's
actually needed, and in this puzzle it was).

The enumerate() method adds counter to the iterable. The returned object
is an enumerate object.

Read more [here](https://www.tutorialspoint.com/enumerate-in-python)