[day05][TIL] simplify search for missing el.
no need to sort the list, loop through the list, find the two consec.
elements which differ by 2, and then returning one of those elements
+/- 1

the puzzle description was purposely misleading. if the whole plain is
full, and our seat is "somewhere in the middle" (ie. not missing from
either from list beginning or end), this is a simple task of finding
who is not there. Thanks Dino!