[day05][TIL] there was no need to split rows/clmns
the formula for getting the ID was 8*row + col. which is the same as
directly converting the whole binary number to decimal. I will not
change it now, but just noting it here (saw it in Dino's solution).

in that case `get_seat_position` becomes redundant, and `seat_ID`
reduces to the single (commented) line.