[day03] refactor and explain how we check slopes
row is the whole string per row, and we find the column (char postion
within that string) based on the given ratio. since we loop through rows
if we have a case that we have a vertical offset greater than 1, the
column postion needs to be divided by this number to maintain the
desired slope.

Not the clearest implementation, I agree...