[day09][TIL] itertools is slower than this
I tried iterating over `itertools.combinations(check_list,2)` but it
turned out to be slower. I guess it is because it creates a new list,
and then needs to access two tuple elements?

Anyway this is just a note that using `itertools` might be nicer, but
this "palenta" approach is faster. At least for this puzzle?