[day09][TIL] hello generators, hello yield!
Instead of creating a new list for every starting position and every
possible list length, we define a function which yields sums for
every length of that list (starting from different starting positions,
as defined in the main loop).

This is a faster approach because there is no need to allocate all these
lists.