[day12] final refactor
the waypoint behaves as ship_direction in many ways, except with the
moving bit. if we take that into account, we can combine most stuff in
a single solve function and then just switch between using the classic
approach and the waypoint approach by using 'part 1'/'part 2' as a
switch.

Not sure if this over-simplified now?