[day19] expand functions to work with example 2:
for example, the tuple of rules is not necessarily always of size 2. I
scaned through my input briefly and it does seem that sizes 1, 2, and 3
should cover all cases.

furhermore, it is possible that two rules simutlaneouosly could be
applied next so I expanded `who to apply next` to return a list of rules
to all be applied in the next go.

next step: parsing the input! (ugh)