[day21] fill dictionary with *all* potential alergens
in my previous implementation that union did not do anything because it
does not change the value in the dictionary!

this causes errors donwstream, so I need to rethink my logic:
I would say that as long as we get `result` we need to eliminate it
from all the potential lists of alergents.