[day20] reduce num of generated side options
any flipped configuration has *ALL* sides reversed. this means that
there is no need to consider "horizontally flipped" and "vertically
flipped" configurations separately. The only difference between the two
is the amount of relative rotation.

For example:

 original:    flipped around X:
 ^---A--->     <---C---^
 |       |     |       |
 D       B     D       B
 |       |     |       |
 <---C---+     +---A--->

 rotated 2x
 original:    flipped around Y:
 ^---C--->     <---C---^
 |       |     |       |
 B       D     D       B
 |       |     |       |
 <---A---+     +---A--->

in other words, if we flip original configuration around X, it is the
same configuration we would arive in by rotating the original config.
twice (obviously direction doesn't matter with rectangles) and flipping
it with respect to Y axis.

This means that in total we have 8 possible configurations per tile.

For the first part we don't care what the configurations are preciselly,
as we just want to compare the sides. To do so we compare the numbers
generated by converting their sides from binary to decimal numbers.