[day20] find "upper left" corner
in order to assemble the first attempt of an image, we need to find a
possible "upper left" corner, ie, the one that matches on its East and
South sides (index positions 1 and 2). I am not sure if every input has
this, but mine does, so I am leaving it as is. Otherwise we would need
to additionally flip and rotate until we got here. So I am thankful for
that.

In the loop we assemble relative rotation and relative flipping of the
matched tile related to the fixed tile. I think this can be used in
assembling the final image, taking into account the configuration of
the matched tile.