[day20] assemble the image...
this is based on many assumptions, namely that there is only one match
possible hence no backtracking needed. out of max two candidates for a
tile we check if it matches either horizontally or vertically and place
it into the assembly.

oriented tiles hold the actual orientations of the tiles and we will use
that to assemble the final image.