[day20] move monster finding to a function
this will not work for a different sequence of transformations?! I need
to flip first, then rotate. otherwise it seems I don't get all the
possible configurations. maybe it has to do something with horizontal
flipping in combination with clockwise rotation? I am to lazy to try and
figure it out now.