[day24] simplify solution/concept using sets
there is no need to store white tiles in the dictionary. it is enough to
have a set of black tiles. the only white tiles that have potential to
become black tiles, are black tiles' neighbours - hence we check in the
expanded mesh. I think this can be further simplified but I will do it
in stages.