[day24][TIL][minblown] hello symmetric difference!
I was certain there must be a better way instead of these conditions,
and of course there is: symmetric difference (^) operator which will
basically add the missing positions or remove them if they are already
present. Since it works on sets, not elements, I needed to create a
single-element set from position like this `{position}`.

Thanks @narimiran !