[day24][TIL] there must be a better way (there is)
instead of copying a set each time and removing elements, why not start
from an empty set and _add_ elements which satisfy a flipped condition?

in that case we are always adding elements to the black set hence the
two conditions can be merged into one.

explanation:
black tiles which have exactly zero, or more than two black adj. tiles
become white tiles. but this same condition can also be rephrased as:
black tiles which have either one or two black ajd. tiles REMAIN black.
hence we can "flip" the condition to choose which tiles "stay" in the
set of black tiles (although effectively they are all added to the new,
output set which starts as emtpy)!

thanks @narimiran !